Move gRPC client (deploy/propose service) to protocol/client folder
